move the visibility listener to common
