+++ app/src/js/components/util.ts
@@ -18,18 +18,3 @@ export function getSubmissionTime (submissions: SubmitData[]) {
   }
   return -1
 }
-
-/**
- * Add callback for the main window visibility change
- */
-export function addVisibilityListener (callback: (visible: boolean) => void) {
-  window.addEventListener('blur', () => {
-    callback(false)
-  })
-  window.addEventListener('focus', () => {
-    callback(true)
-  })
-  window.addEventListener('visibilitychange', () => {
-    callback(!document.hidden)
-  })
-}
